requirements.txt
azure-functions
jinja2